Simplifying etcd TLS configuration/use
