+++ monitoring/healthz.go
@@ -32,10 +32,11 @@ const healthzTLSHandshakeTimeout = 10 * time.Second
 const dialTimeout = 30 * time.Second
 const keepAlivePeriod = 30 * time.Second
 
-// HttpResponseChecker is a function that can validate service status from response
+// HttpResponseChecker is a function that can validate service health
+// from the provided response
 type HttpResponseChecker func(response io.Reader) error
 
-// HttpHealthzChecker is a health checker that can check servhice status over HTTP
+// HttpHealthzChecker is a health.Checker that can validate service health over HTTP
 type HttpHealthzChecker struct {
 	name    string
 	URL     string
@@ -74,34 +75,30 @@ func (r *HttpHealthzChecker) Check(reporter health.Reporter) {
 	})
 }
 
+// NewHTTPHealthzChecker creates a health.Checker for an HTTP health endpoint
+// using the specified URL and a custom response checker
 func NewHTTPHealthzChecker(name, URL string, checker HttpResponseChecker) health.Checker {
 	defaultTransport := http.RoundTripper(nil)
 	return NewHTTPHealthzCheckerWithTransport(name, URL, defaultTransport, checker)
 }
 
+// NewUnixSocketHealthzChecker returns a new Checker that tests
+// the specified unix domain socket path and URL
+func NewUnixSocketHealthzChecker(name, URL, socketPath string, checker HttpResponseChecker) health.Checker {
+	transport := &http.Transport{
+		Dial: func(network, addr string) (net.Conn, error) {
+			return net.Dial("unix", socketPath)
+		},
+	}
+	return NewHTTPHealthzCheckerWithTransport(name, URL, transport, checker)
+}
+
+// NewHTTPHealthzCheckerWithTransport creates a health.Checker for an HTTP health endpoint
+// using the specified transport, URL and a custom response checker
 func NewHTTPHealthzCheckerWithTransport(name, URL string, transport http.RoundTripper, checker HttpResponseChecker) health.Checker {
 	client := &http.Client{
 		Transport: transport,
 		Timeout:   healthzCheckTimeout,
-	}
-	return &HttpHealthzChecker{
-		name:    name,
-		URL:     URL,
-		client:  client,
-		checker: checker,
-	}
-}
-
-// NewUnixSocketHealthzChecker returns a new Checker that tests
-// the specified unix domain socket path and URL
-func NewUnixSocketHealthzChecker(name, URL, socketPath string, checker HttpResponseChecker) health.Checker {
-	client := &http.Client{
-		Timeout: healthzCheckTimeout,
-		Transport: &http.Transport{
-			Dial: func(network, addr string) (net.Conn, error) {
-				return net.Dial("unix", socketPath)
-			},
-		},
 	}
 	return &HttpHealthzChecker{
 		name:    name,
